rename package and enable scheduling in bot-module
