+++ bot/src/main/java/ru/tinkoff/edu/java/bot/BotApplication.java
@@ -3,12 +3,14 @@ package ru.tinkoff.edu.java.bot;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import ru.tinkoff.edu.java.bot.configuration.ApplicationConfig;
 import ru.tinkoff.edu.java.bot.model.core.BotCreator;
 import ru.tinkoff.edu.java.bot.model.command.*;
 import ru.tinkoff.edu.java.bot.scrapper.api.ScrapperClient;
 
 @SpringBootApplication
+@EnableScheduling
 @EnableConfigurationProperties(ApplicationConfig.class)
 public class BotApplication {
         public static void main(String[] args) {
@@ -23,7 +25,5 @@ public class BotApplication {
                         new TrackCommand(client),
                         new UntrackCommand(client),
                         new HelpCommand());
-
-
         }
 }
